refactor: component TodoList was completely migrated from the App.tsx
